rename "internal" to "lib" so this library plays nicely with vendoring
