+++ .travis.yml
@@ -32,7 +32,7 @@ install:
 - cd ../..
 - pwd
 script:
-- ginkgo -cover -race -r -keepGoing -succinct -randomizeSuites internal/lua
+- ginkgo -cover -race -r -keepGoing -succinct -randomizeSuites lib/lua
 - ginkgo -cover -race -r -keepGoing -succinct -randomizeSuites types
 - ginkgo -cover -race -r -keepGoing -succinct -randomizeSuites pkg
 - find tools -name "*.go" -type f -print0 | xargs -0 -n1 go build
