+++ lib/lua/instance.go
@@ -15,7 +15,7 @@
 package lua
 
 import (
-	"github.com/aerospike/aerospike-client-go/internal/lua/resources"
+	"github.com/aerospike/aerospike-client-go/lib/lua/resources"
 	"github.com/aerospike/aerospike-client-go/logger"
 	"github.com/aerospike/aerospike-client-go/types"
 	"github.com/yuin/gopher-lua"
